Changed the description (flavor text) in several AI decks. the old descriptions consisted of flavor texts of existing cards. Even though the flavor texts are often just a few words long, I'm pretty certain that this is copyrighted material, and before we go to great lengths to find out whether any interpretation of "fair use" would allow us to use them nevertheless, I think it's better to just remove them and come up with texts that we don't have to worry about.
Note: I used Google to match our deck descriptions with existing card flavor texts, and then used Google again to check whether the text was in the public domain (the fact that a verse from the Rime of the Ancient Mariner appears on the Scathe Zombies card, does obviously not preclude us from using Coleridge's work ourselves). I may have missed something in wither process, please report if I did.

For the time being, this is unfortunately a loss in quality, since the official flavor texts were obviously better than the replacement texts I thought up in a few minutes. Feel free to replace my texts with something better. This could be original work of yours, or something that's clearly in the public domain. The latter includes classical literature and poetry btw. We already have Milton, Browning, and Coleridge in the flavor texts, and we could certainly use more of that. Unfortunately I've read about 100 SciFi/Fantasy books for every book of classical poetry I even touched, so I'm in a bad position to provide such material, but perhaps someone else can.

I discussed this change with wololo beforehand, because it required changes to decks that weren't made by me.

Please review / criticize.

git-svn-id: http://wagic.googlecode.com/svn/trunk@1179 01922e74-a4c3-11dd-a5cd-697bfd0a0cef