Psyringe - added test suite scripts for issues i163 to i168.
Also added two scripts that don't demonstrate bugs (apparently the bug reports that I used to write the scripts were either false or outdated). nevertheless it doesn't hurt to have these scripts in the test suite, it prevents us from breaking something for the respective cards in the future and not noticing it.

The script fear_i147.txt has been moved into the "generic" folder and added to the regular test sequence for the same reason.

Also fixed the script generic/tokens.txt. It was referencing The Hive once via name and once via ID. This caused problems when (unofficial) Beta edition _cards.dat was installed.

git-svn-id: https://wagic.googlecode.com/svn/trunk@1247 01922e74-a4c3-11dd-a5cd-697bfd0a0cef