Changed the code of all cards with optional cycling effects (ONS).
EXAMPLE: The older code of Death Pulse was:

autohand={1}{B}{B}:cycling && may -1/-1 target(creature)

I tested this ingame and the following happened: 
1. I pay the cycling cost. 
2. The engine gives me the option to target a creature. 
3. I choose a creature.
4. The cycling effect resolves: Death Pulse goes to the graveyard and I draw a card, but the targeted creature does not get the p/t malus.

This is seems to be because the cycling effect is coded as  option ("may") and the cycle process (dicard this:Draw:1) is coded as a "must".

I found a new, successfully tested way to code such cards by giving them 2 autohand-lines instead of only one: 
The first one is without optional effect and the second one is with optional effect.
Ingame, if you cycle Death Pulse f.e., a window will appear to choose between those two auto-lines. WORKS GREAT AND DOES NOT BREAK THE CARD!! :)  

   

git-svn-id: https://wagic.googlecode.com/svn/trunk@1492 01922e74-a4c3-11dd-a5cd-697bfd0a0cef