Psyringe - fixed Mirri the Cursed (PLC). Temporarily fixed Reiver Demon (MRD). Thanks to Raphael_Carlo for the bug report.
Important: I just realized that the fix to Reiver Demon will probably lead to problems later on. The problem is this: Reiver Demon has an ability "When yu play this from your hand ...". This was coded as @moveto(this|mybattlefield) from(mayhand). Which doesn't work, because the card doesn't actually enter the battlefield from the hand, it enters from the stack. So, "@moveto(this|mybattlefield) from(maystack)" has the desired effect. However, IIUC, spells played from anywhere else (e.g. from the graveyard, if they allow that) should be placed on the stack to. Hence, the "fixed" trigger will also fire in this case, which is wrong.

This means that we can't actually code "played from ..." entirely correct at the moment. Is there a solution for this?

According to the rules, only a copy of a spell enters the stack (never the spell itself), so could the "from()" part access where the card actually came from?

Do we need an @played trigger to differentiate between moving a card and actually playing it?

git-svn-id: https://wagic.googlecode.com/svn/trunk@1659 01922e74-a4c3-11dd-a5cd-697bfd0a0cef