J : * Fixes for compilation on 64-bit architectures :   string position-returning functions return size_t, not uint.   Fix that. * Fixes warnings with new gcc about non-fixed string AND no arguments.
NOTE : 64-bit architectures still do not compile. There are 2 different places
where printf is used with a specifier that is not large enough on 64-bit because
size_t is now a ulong and not a uint. The solution on GNU systems is the %zu
specifier, but as I'm not sure it is supported on windows I don't want to
introduce a huge hard-to-notice bug, so I'll do it at a time when I got a
windowser to back me up.



git-svn-id: http://wagic.googlecode.com/svn/trunk@1742 01922e74-a4c3-11dd-a5cd-697bfd0a0cef