Psyringe - Checked the manacost, color, type, subtype, power, toughness, and loyalty data of all cards in mtg.txt and mtg_todo.dat, and did about 200 bugfixes, optimizations, and cosmetic changes. All cards in both files should now have correct values for the keywords mentioned above.
Note: I checked all changes manually, so in the cases where the value *has* to be different for the card to work (e.g.Covetous Dragon), I left the values as they were. In these cases I added a comment to the card code which notes why the values need to be different, so that no one else breaks the card by "correcting" them.

Note 2: You may have noticed that I used a fixed sequence of lines for all cards. That was a byproduct of the file unification process, but I think it's also rather useful. Above the "text=" line are the lines that we need to code (abilities=, target=, auto= etc.). Below the "text=" line are the fixed values (mana=, type=, power= etc.) which we only need to touch if either the card gets errata'd, or we need to do some tricky coding (e.g. Covetous Dragon). So, in other words, everything below the "text=" line should be okay now in both files, and doesn't need to be checked when implementing or debugging a card.

git-svn-id: https://wagic.googlecode.com/svn/trunk@1823 01922e74-a4c3-11dd-a5cd-697bfd0a0cef