Fix for the Snap card not triggering the 2nd untap.  Basically, stole the method from GrimmHatter's post:
http://www.wololo.net/forum/viewtopic.php?f=21&t=1430&start=40#p13465

Seems to work from my testing so far.  If this is the preferred route, we should probably sweep through & track down all the other cards that use the token counter hack, as that method used to work in 0.12.1 and broke some time after that.

(Somebody please sanity check this for me, as this is my first foray in messing with card rules.)

git-svn-id: https://wagic.googlecode.com/svn/trunk@2392 01922e74-a4c3-11dd-a5cd-697bfd0a0cef