Follow up on this comment from Xawotihs:
"It's currently unclear for me why several instances of the same hgeparticlesystem with the same .psi file are created..."

The GameApp::Create() call was initializing an array of particles, but it wasn't actually stashing them in the psi cache.  The MenuItems also create the particles, but these are stashed correctly in the cache at that point.  So, I simply removed the unused particle array loading code since it wasn't actually caching anything correctly & therefore simply adding redundant reading of the same particle files.


git-svn-id: https://wagic.googlecode.com/svn/trunk@2399 01922e74-a4c3-11dd-a5cd-697bfd0a0cef