seperated a varible that determined the amount of sliding of cards based on how many noncreatures and nonland cards you had.
what this change does is this. if i have 20 enchantments on the battlefield Ai's creatures and lands no longer slide to the right. and vice versa. made no sense that if the other player had no artifacts/enchantments that HIS lands and creatures would slide simply because the OTHER player had artifacts/enchantments.

git-svn-id: http://wagic.googlecode.com/svn/trunk@2592 01922e74-a4c3-11dd-a5cd-697bfd0a0cef