PC specific change (ie PSP not impacted) - OpenGL is not thread safe, only one thread is allowed to have the context at any given time.  The JGE library was not written with multithreading in mind, so the LoadTexture call doesn't work when called on a worker thread.  In order to get my threading changes to work, I had to split out the OpenGL calls to happen on a point where we're guaranteed to be back on the main thread.
Functionally, this has no impact to the current build - all I've done is effectively the texture loading to happen in JQuad's constructor instead of inside the LoadTexture() function.

Side note - I probably broke the IOS build with this change.  Sorry, but I can only build so many targets at a time - I'm simply not set up to build this platform.  As it's not a shipping target, I have to consider it a 2nd class citizen...


git-svn-id: https://wagic.googlecode.com/svn/trunk@2714 01922e74-a4c3-11dd-a5cd-697bfd0a0cef