Fix for [PSP|PC]ability menus and card view switch causes a crash.  This broke as a side effect to the addition of the 'details' option that Mike added - the ActionStack had very rigid interpretations of what a valid control ID would be (either -1, or a valid id in its container).  As kInfoMenuID is assigned a value of -200, this would crash when attempting to look up that index value in the vector.
I've patched the problem so that ActionLayer checks for a valid ID before doing anything, and does a no-op for IDs it doesn't know how to handle. However, this only fixes the problem during gameplay.  It's quite possible that hitting the triangle button in other places in the app might equally cause a crash.


Issue: 544


git-svn-id: http://wagic.googlecode.com/svn/trunk@2741 01922e74-a4c3-11dd-a5cd-697bfd0a0cef