corrected the following cards.... Lake of the Dead trigon of mending lord of shatterskull pass gorilla shaman kumano nightsky mimic
this one is debatable...lovisa cold eyes
what i did here was code lovisa EXACTLY as written on her card...
the erratta that was given to this card becuase it was considered "overpowered" was redicoulous....absolutely redicoulous...by the card her code is exactly as follows...
this card is FAR to basic to say there is a mistake in my understanding of it.
subtype=human lord <---this is on the card...
she is NOT a barbarian...
exact wording on the card
Warriors, Berserkers, and Barbarians creatures get +2/+2 and have haste.
this is coded

auto=lord(warrior) 2/2 
auto=lord(berserker) 2/2 
auto=lord(barbarian) 2/2 
auto=lord(warrior) haste 
auto=lord(berserker) haste
auto=lord(barbarian) haste

this is the real lovisa....
errattas are fine, but its wotc that made the mistake of creating this card in the first place :) if you hit gatherer i think you will agree...
http://gatherer.wizards.com/Pages/Card/Details.aspx?multiverseid=113542
following the actual card on this one...ditching the erratta.

git-svn-id: https://wagic.googlecode.com/svn/trunk@3154 01922e74-a4c3-11dd-a5cd-697bfd0a0cef