updated thallid test to look for the correct token id...this is why you should not add multiple tokens for the same creature as entries in card.dat...thallid looks for -1924...the original id....-11492115 is a new card.dat info which was added and i complained about the constant fails in the test suite from these types of additions.
this was intruduced as a workaround to force a card to be able to display muliple types of tokens, which wagic can support with this workaround but causes fails in test suite....

Sarpadian Empires, Vol. VII

this is the card that created this fail.

in the future we need to be careful about adding tokens to card.dat...there are 2 different systems for getting a token to appear for a card. lets use what we have.

git-svn-id: https://wagic.googlecode.com/svn/trunk@3272 01922e74-a4c3-11dd-a5cd-697bfd0a0cef