code change for life and limb, auto=lord(*[saproling;forest]) transforms((saproling forest creature land,setpower=1,settoughness=1,newability[{t}:add{g}]))
this is close enough for me, its 99% correct in my books, you can declare either forest and saprolings as attackers, target them as thier respective new types, the only thing that is not correct about it is that they can tap for mana the turn they come into play, however they have summoning sickness and can not attack.

im am not going to change the attack/block rules or manaproducer reactToClick just to satisfy this bug in the only card in MTG that does this. 

git-svn-id: https://wagic.googlecode.com/svn/trunk@3326 01922e74-a4c3-11dd-a5cd-697bfd0a0cef