First pass at reducing the overall memory footprint:  moved GetFormattedText() out of CardPrimitives and into MTGCard.  The idea being, only keep the formatted text around for cards that are actually in use.
Also fixed a subtle memory pooling issue in the RenderCountersBig() routine:  if you're in regular card display mode, the idea is that formatted text is only fetched if you flip into alternate render mode.  However, in this function, if counters are being drawn, it would fetch the formatted text in order to determine where to draw the counters, EVEN IF the counters count was zero.  So it had nothing to draw, but it meanwhile pooled the formatted strings into memory anyway.


git-svn-id: https://wagic.googlecode.com/svn/trunk@3518 01922e74-a4c3-11dd-a5cd-697bfd0a0cef