3 fixes, first you are not allowed to mulligen the opponents hand.
second persist was using the stack on its returning effect, this is incorrect as per MTG rules for persist, the card is supposed to go from the grave directly to inplay(grave -> temp -> inplay, in our case).

third, tokens models don't have names, so i make it check now for a name size on the model before clearing and resetting in tranformer.


git-svn-id: https://wagic.googlecode.com/svn/trunk@3669 01922e74-a4c3-11dd-a5cd-697bfd0a0cef