added the support to fix avatar of might...you can now do otherrestriction=type(creature|mybattlefield)~lessthan~type(creature|opponentbattlefield)-3
which means, if the statement above is true you can play the card....if you subtract 3 from the amount the opponent has in creatures, and you are still less then the opponent...then that means he has atleast 4 more creatures than you do.

so right after the type(blah) you can add +number or -number...and it will modify the amount to compare.

git-svn-id: https://wagic.googlecode.com/svn/trunk@3767 01922e74-a4c3-11dd-a5cd-697bfd0a0cef