fixed a bug in targetchoosing where a player would get stuck in a targetchooser under certain situations as follows: opponent play an ability, player interupts and clicks a card with TargetChooser which has no valid targets; player then tries to cancel targetchooser, instead the stack is displayed again, player is then locked in a TargetChooser unable to click other cards.
originally i was going to do it in the reactToClick however, after thinking about it, it made no sense to me that a targetchooser would be set and a cardwaitingfortargets would be set if you know ahead of time that no valid targets existed.

now when clicking cards which have TargetChoosers and no valid targets, no tc is set, and no cardiswaitingfortargets; instead return 0.

git-svn-id: http://wagic.googlecode.com/svn/trunk@3930 01922e74-a4c3-11dd-a5cd-697bfd0a0cef